[CSL-1165] Move BatchOp stuff to a separate module
