+++ src/Pos/Explorer/DB.hs
@@ -19,8 +19,8 @@ import           Pos.Binary.Class      (encodeStrict)
 import           Pos.Context.Functions (GenesisUtxo, genesisUtxoM)
 import           Pos.Core              (unsafeAddCoin)
 import           Pos.Core.Types        (Address, Coin)
-import           Pos.DB.Class          (MonadRealDB, MonadDBRead, getGStateDB)
-import           Pos.DB.Functions      (RocksBatchOp (..), rocksGetBytes)
+import           Pos.DB                (MonadDBRead, MonadRealDB, RocksBatchOp (..),
+                                        getGStateDB, rocksGetBytes)
 import           Pos.DB.GState.Common  (gsGetBi, gsPutBi, writeBatchGState)
 import           Pos.Explorer.Core     (AddrHistory, TxExtra (..))
 import           Pos.Txp.Core          (TxId, TxOutAux (..), _TxOut)
